chore: refactor websocket layer
