+++ AGENTS.md
@@ -1,4 +1,4 @@
-## Nächste Aufgaben (Sprint 1)
-1. [P0] WebSocket-Handshake unter `/ws` mit JWT-Authentifizierung implementieren.
-2. [P1] Tests für WebSocket-Handshake aktualisieren und erweitern.
-3. [P1] Dokumentation des WebSocket-Endpunkts ergänzen und change.log aktualisieren.
+## Nächste Aufgaben (Sprint 2)
+1. [P0] Identifiziere und entferne die gesamte von Gemini generierte WebSocket-Legacy-Implementierung im Frontend und Backend. Dokumentiere den Vorgang im change.log.
+2. [P0] Entwerfe das neue, kanalisierte WebSocket-Datenmodell und das JSON-Nachrichtenprotokoll. Dokumentiere die Spezifikationen detailliert in backend.md.
+3. [P1] Implementiere den serverseitigen WebSocket-Endpunkt inklusive des Token-basierten Authentifizierungs-Checks beim Handshake-Prozess.
