+++ test/server.test.js
@@ -129,41 +129,6 @@ test('GET /api/users returns list of users', async () => {
   await new Promise(r => server.close(r));
 });
 
-test('POST /api/hive/log broadcasts to websocket clients', async () => {
-  const server = await start();
-  const port = server.address().port;
-
-  const reg = await fetch(`http://localhost:${port}/api/auth/register`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json' },
-    body: JSON.stringify({ username: 'eve', email: 'e@e.com', password: 'p' })
-  });
-  const { token } = await reg.json();
-
-  const ws = new WebSocket(`ws://localhost:${port}/ws?token=${token}`);
-  await new Promise(res => ws.on('open', res));
-
-  const logPromise = new Promise(resolve => {
-    ws.on('message', data => resolve(JSON.parse(data)));
-  });
-
-  await fetch(`http://localhost:${port}/api/hive/log`, {
-    method: 'POST',
-    headers: { 'Content-Type': 'application/json', Authorization: `Bearer ${token}` },
-    body: JSON.stringify({ message: 'hello' })
-  });
-
-  const msg = await logPromise;
-  assert.strictEqual(msg.event, 'hive-log');
-  assert.strictEqual(msg.data.message, 'hello');
-
-  const listRes = await fetch(`http://localhost:${port}/api/hive/logs`);
-  const list = await listRes.json();
-  assert.ok(list.some(l => l.message === 'hello'));
-
-  ws.close();
-  await new Promise(r => server.close(r));
-});
 
 test('GET /api/hive/logs returns persisted logs', async () => {
   const server = await start();
